feat: unwrap account namespace to flat exports + barrel
